+++ src/pages/Functional/CreateVote/_components/VoteForm/VoteForm.tsx
@@ -7,6 +7,7 @@ import * as S from "../../../_components/Functional.styles";
 import CalendarIcon from "../../../../../assets/icons/Calender.svg?react";
 import Card from "../../../../../components/Card";
 import VoteTimes from "./VoteTimes";
+import PriceInput from "./PriceInput";
 
 interface VoteFormProps {
   tripPlanType: "COURSE" | "SCHEDULE";
@@ -19,6 +20,15 @@ interface VoteFormProps {
 export default function VoteForm({ tripPlanType, roomId, onCalendar, onSearch, selectedPlace }: VoteFormProps) {
   const { control, watch, setValue } = useVoteForm(tripPlanType, roomId);
   const isSchedule = tripPlanType === "SCHEDULE";
+
+  useEffect(() => {
+    const storedStartDate = localStorage.getItem("voteForm_startDate");
+    const storedEndDate = localStorage.getItem("voteForm_endDate");
+
+    if (storedStartDate) setValue("startDate", storedStartDate);
+    if (storedEndDate) setValue("endDate", storedEndDate);
+  }, [setValue]);
+
   const startDate = watch("startDate");
   const endDate = watch("endDate");
 
@@ -35,13 +45,6 @@ export default function VoteForm({ tripPlanType, roomId, onCalendar, onSearch, s
       setValue("location", selectedPlace);
     }
   }, [selectedPlace, setValue]);
-
-  const formatPrice = (price?: number) => {
-    if (!price || price === 0) return "20만원";
-    const mainUnit = Math.floor(price / 10000);
-    const subUnit = price % 10000;
-    return subUnit === 0 ? `${mainUnit}만 원` : `${mainUnit}만 ${subUnit}원`;
-  };
 
   return (
     <Card>
@@ -66,42 +69,7 @@ export default function VoteForm({ tripPlanType, roomId, onCalendar, onSearch, s
       {isSchedule && (
         <>
           <Card.Item label="가격">
-            <Controller
-              name="scheduleDetails.price"
-              control={control}
-              render={({ field }) => {
-                let priceValue = field.value ? field.value.replace(/\D/g, "") : "";
-                let priceNumber = priceValue ? parseInt(priceValue, 10) : 0;
-
-                if (priceNumber > 10000000 || priceValue.length > 8) {
-                  priceNumber = 10000000;
-                }
-
-                return (
-                  <S.Input
-                    type="text"
-                    placeholder={`${nights}박 / 20만원`}
-                    value={field.value || ""}
-                    onChange={(e) => {
-                      let newValue = e.target.value.replace(/\D/g, "").slice(0, 7);
-                      field.onChange(newValue);
-                    }}
-                    onBlur={() => {
-                      if (field.value) {
-                        let parsedValue = parseInt(field.value, 10) || 0;
-                        field.onChange(`${nights}박 / ${formatPrice(parsedValue)}`);
-                      }
-                    }}
-                    onKeyDown={(e) => {
-                      if (e.key === "Backspace") {
-                        let newValue = field.value ? field.value.slice(0, -1) : "";
-                        field.onChange(newValue);
-                      }
-                    }}
-                  />
-                );
-              }}
-            />
+            <PriceInput control={control} nights={nights} />
           </Card.Item>
 
           <S.StyledDivider />
@@ -109,7 +77,9 @@ export default function VoteForm({ tripPlanType, roomId, onCalendar, onSearch, s
       )}
 
       <S.StyledCardItem>
-        <span className="text">날짜 {startDate || "미정"} ~ {endDate || "미정"}</span>
+        <span className="text">
+          날짜 {startDate ? startDate : "미정"} ~ {endDate ? endDate : "미정"}
+        </span>
         <S.IconWrapper onClick={onCalendar} className="icon">
           <CalendarIcon />
         </S.IconWrapper>
